specs fixed on core

+++ app/models/order/verbal.rb
@@ -304,10 +304,12 @@ module Order
     end
 
     def remove_busy_days
-      reservation_dates.confirmed.each do |date|
-        assignee.busy_days.where(date: date.date).delete_all
+      if assignee.present?
+        reservation_dates.confirmed.each do |date|
+          assignee.busy_days.where(date: date.date).delete_all
+        end
+        assignee.save
       end
-      assignee.save
     end
 
     def update_notification
